ML-8 - Header design + responsive
