+++ src/app/app.component.html
@@ -1,3 +1,3 @@
-<app-cart [cart]="cart"></app-cart>
+<app-header></app-header>
 <app-nav-bar></app-nav-bar>
 <router-outlet></router-outlet>
