+++ src/app/app.component.scss
@@ -32,8 +32,5 @@
     display: none;
 }
 
-.dialog-active {
-}
-
 @media only screen and(max-width: 600px) {
 }
