+++ src/app/app.component.ts
@@ -10,24 +10,6 @@ import * as _ from 'lodash';
 })
 export class AppComponent {
   title = 'recipe-tutorials';
-  cart: any = new Object(null);
 
-  constructor(private firebaseService: FirebaseService) { }
-
-  ngOnInit() {
-    this.firebaseService.getCartInfo().valueChanges().subscribe(cart => {
-      if (cart) {
-        this.cart = new Object(cart);
-        let counts = {};
-        let formattedArr = [];
-        //Counting repetitions of items
-        this.cart.cartItems.forEach((x) => { counts[x.toLowerCase()] = (counts[x.toLowerCase()] || 0) + 1; });
-        for (let prop in counts) {
-          formattedArr.push(`${prop} ${counts[prop]}x`);
-        }
-        this.cart.cartItems = formattedArr;
-      }
-    })
-  }
 
 }
