+++ src/app/header/header.component.html
@@ -1,0 +1,13 @@
+<div class="header-container">
+    <div class="logo-container">
+        <img src="../../assets/logo.png" alt="Site Logo" class="logo-img">
+    </div>
+
+    <div class="search-bar-container">
+        <input class="search-input" name="search" type="text" placeholder="Search">
+        <button class="search-btn">
+            <img class="search-icon" src="../../assets/search.svg" alt="Search Icon">
+        </button>
+    </div>
+    <app-cart [cart]="cart"></app-cart>
+</div>
